pyproject.toml
[tool.poetry]
name = "copilot-automotive"
version = "0.1.0"
description = ""
authors = ["aRaiko.Funakami <raiko.funakami@access-company.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
pydantic = "^2.5.2"
python = "^3.11"
openai = "^1.16.1"
flask = "^3.0.2"
langid = "^1.1.6"
selenium = "^4.19.0"
webdriver-manager = "^4.0.1"
ppad = "^0.0.11"
pure-python-adb = "^0.3.0.dev0"
langchain-openai = "^0.2.9"
langchain-community = "^0.3.8"
langgraph = "^0.2.53"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
